Add nightly feature
